requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
certifi==2018.1.18
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
decorator==4.1.2
defusedxml==0.5.0
Django==1.11
django-allauth==0.35.0
django-extensions==1.9.8
giphypop==0.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
msgpack-python==0.4.8
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
zope.interface==4.4.3
